pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backpropagate"
version = "0.1.0"
description = "Production-ready headless LLM fine-tuning with smart defaults, Windows support, and modular architecture"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Mikey Frilot" }
]
keywords = [
    "llm",
    "fine-tuning",
    "lora",
    "qlora",
    "unsloth",
    "training",
    "machine-learning",
    "headless",
    "api",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

# Core dependencies - minimal for basic functionality
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "datasets>=2.14.0",
    "trl>=0.7.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "tenacity>=8.0.0",
    "packaging>=21.0",  # Version parsing for security checks
]

[project.optional-dependencies]
# Individual feature groups - install only what you need
unsloth = [
    "unsloth>=2024.1",  # Unsloth for 2x faster training
]
ui = [
    "gradio>=5.6.0",  # Web UI - minimum 5.6.0 for CVE-2025-23042 fix
]
validation = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",  # Config validation
]
export = [
    "llama-cpp-python>=0.2.0",  # GGUF export
]
monitoring = [
    "psutil>=5.9.0",  # System monitoring
    "wandb>=0.15.0",  # Weights & Biases logging
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
logging = [
    "structlog>=24.1.0",  # Structured logging (2026 best practices)
]
security = [
    "PyJWT>=2.8.0",  # JWT authentication
    "cryptography>=42.0.0",  # Secure token generation
]

# Bundle presets for convenience
standard = [
    "backpropagate[unsloth,ui]",  # Recommended for most users
]
full = [
    "backpropagate[unsloth,ui,validation,export,monitoring,observability,logging,security]",  # Everything
]
production = [
    "backpropagate[unsloth,ui,validation,logging,security]",  # Production deployment
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-xdist>=3.5.0",     # Parallel test execution
    "pytest-timeout>=2.3.0",   # Test timeouts
    "hypothesis>=6.100.0",     # Property-based testing
    "mutmut>=2.4.0",           # Mutation testing
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    # Security scanning tools
    "bandit>=1.7.0",           # Security linter
    "pip-audit>=2.7.0",        # Dependency vulnerability scanner
    "safety>=2.3.0",           # Dependency security checker
    "packaging>=21.0",         # Version parsing for security checks
]

[project.urls]
Homepage = "https://github.com/mikeyfrilot/backpropagate"
Documentation = "https://github.com/mikeyfrilot/backpropagate#readme"
Repository = "https://github.com/mikeyfrilot/backpropagate"
Issues = "https://github.com/mikeyfrilot/backpropagate/issues"

[project.scripts]
backpropagate = "backpropagate.cli:main"
backprop = "backpropagate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["backpropagate"]

[tool.hatch.build.targets.sdist]
include = [
    "/backpropagate",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
    "C901",   # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__
"tests/*" = ["ARG"]       # unused arguments in tests

[tool.ruff.isort]
known-first-party = ["backpropagate"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--timeout=60",
]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests that require GPU or external services",
    "gpu: marks tests that require a GPU",
    "hypothesis: marks property-based tests",
]
filterwarnings = [
    # Ignore deprecation warning about TRANSFORMERS_CACHE - this is user's env config
    "ignore:Using `TRANSFORMERS_CACHE` is deprecated:FutureWarning",
]

[tool.coverage.run]
source = ["backpropagate"]
branch = true
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
fail_under = 50
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = [
    "B101",  # assert - fine in non-production code
    "B104",  # bind 0.0.0.0 - expected for UI server, users should firewall
    "B404",  # subprocess import - required for GGUF/Ollama export
    "B603",  # subprocess call - inputs are validated in export.py
    "B607",  # partial path - ollama/llama.cpp are user-installed CLIs
    "B615",  # HuggingFace downloads - model names are user-provided, not a vulnerability
]

[tool.bandit.assert_used]
skips = ["*_test.py", "test_*.py", "conftest.py"]

[tool.mutmut]
paths_to_mutate = "backpropagate/"
tests_dir = "tests/"
runner = "python -m pytest -x --tb=no -q"
dict_synonyms = "Struct, NamedTuple"

[tool.poodle]
source_folders = ["backpropagate"]
[tool.poodle.runner_opts]
command_line = "pytest F:/AI/backpropagate/tests/ -x --assert=plain -o 'pythonpath={PYTHONPATH}' --ignore=F:/AI/backpropagate/tests/test_cli.py --ignore=F:/AI/backpropagate/tests/test_cli_extended.py"
